+++ src/backend/storage/file/fd.c
@@ -91,7 +91,7 @@
 #define VFD_CLOSED (-1)
 
 #define FileIsValid(file) \
-	((file) > 0 && (file) < SizeVfdCache && VfdCache[file].fileName != NULL)
+	((file) > 0 && (file) < (int) SizeVfdCache && VfdCache[file].fileName != NULL)
 
 #define FileIsNotOpen(file) (VfdCache[file].fd == VFD_CLOSED)
